Reverted Changes and put types into proper robot types file
