+++ public/usd-viewer/hydra/ThreeJsRenderDelegate.js
@@ -29,13 +29,9 @@ export class ThreeRenderDelegateInterface {
    * @returns
    */
   createRPrim(typeId, id, instancerId) {
-    if (DEBUG_PRIMS) console.log("Creating RPrim:", typeId, id, instancerId);
+    if (DEBUG_PRIMS) console.log("Creating RPrim:", typeId, id);
     const mesh = new HydraMesh(id, this);
-    // Record instancing information so downstream can distinguish prototypes vs instances
-    mesh._isInstance = !!instancerId;
-    mesh._instancerId = instancerId;
     this.meshes[id] = mesh;
-    // Removed verbose RPrim creation logs
     return mesh;
   }
 
@@ -49,28 +45,13 @@ export class ThreeRenderDelegateInterface {
     if (t === "material") {
       const material = new HydraMaterial(id, this);
       this.materials[id] = material;
-      if (this?.config?.debugTransforms) {
-        try {
-          console.debug(`[USD] SPrim(material) created: id=${id}`);
-        } catch {}
-      }
       return material;
     }
     // Acknowledge camera and light sprims to prevent hydra warnings
     if (t === "camera") {
-      if (this?.config?.debugTransforms) {
-        try {
-          console.debug(`[USD] SPrim(camera) created: id=${id}`);
-        } catch {}
-      }
       return new HydraCamera(id, this);
     }
     if (t.includes("light")) {
-      if (this?.config?.debugTransforms) {
-        try {
-          console.debug(`[USD] SPrim(light) created: id=${id}`);
-        } catch {}
-      }
       return new HydraLight(id, this);
     }
   }
